+++ mrz-3-edition/utils/cramerSolver.m
@@ -1,0 +1,34 @@
+function [x] = cramerSolver(A, b)
+% CRAMERSOLVER Solves linear system by Cramer's rule.
+%
+% [x] = cramerSolver(A, b)
+%
+% Input:
+% A - matrix of coefficients
+% b - known terms
+%
+% Output:
+% x - solution of Ax = b
+
+% Copyright 2017 Stefano Fogarollo
+%
+% Licensed under the Apache License, Version 2.0 (the "License");
+% you may not use this file except in compliance with the License.
+% You may obtain a copy of the License at
+%
+% http://www.apache.org/licenses/LICENSE-2.0
+%
+% Unless required by applicable law or agreed to in writing, software
+% distributed under the License is distributed on an "AS IS" BASIS,
+% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+% See the License for the specific language governing permissions and
+% limitations under the License.
+% TODO: move to right chapter
+
+n = size(A, 1);  % number of rows (= number of columns)
+for j = 1 : n
+   B = A;
+   B(:, j) = b;
+   x(j) = det(B) / det(A);
+end
+x = x';